requirements.txt
psycopg2
Django
django-bootstrap5
crispy-bootstrap5
djangorestframework
jinja2
kubernetes
requests
